requirements_classification.txt
absl-py==0.10.0
astor==0.8.1
astunparse==1.6.3
attrs==20.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.6.0
conda==4.3.16
configparser==5.0.0
cycler==0.10.0
dask==2.25.0
decorator==4.4.2
docopt==0.6.2
gast==0.2.2
glob2==0.5
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.25.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
iniconfig==1.1.1
install==1.3.4
joblib==0.17.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.32.1
logger==1.4
Markdown==3.2.2
matplotlib==3.2.1
more-itertools==8.5.0
networkx==2.5
num2words==0.5.10
numba==0.49.1
numpy==1.18.4
oauthlib==3.1.0
opencv==1.0.1
opencv-python==3.4.4.19
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.3
pickle5==0.0.11
Pillow==7.2.0
pluggy==0.13.1
protobuf==3.13.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pyocclient==0.6
pyparsing==2.4.7
pytest==6.2.1
pytest-gc==0.0.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-image==0.14.2
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.11.1
Shapely==1.7.0
six==1.15.0
sklearn==0.0
tbb==2021.1.1
tensorboard==2.1.1
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.9.0
tqdm==4.48.2
umap-learn==0.4.2
unrar==0.4
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.3.7
zipp==3.1.0
